+++ src/commands.ts
@@ -96,9 +96,9 @@ export function getSortCommandText(toRunInTerminal: boolean, useProjectSpecific:
     }
 
     const optionalArgs = addOptionalArgs ? ' $*' : '';
-    let extraOptions = ' ' + getOverrideConfigByPriority([folderKey, 'default', ''], 'extraOptions', true).trimRight();
+    let extraOptions = ' ' + getOverrideConfigByPriority([folderKey, '', 'default'], 'extraOptions', true).trimRight();
     extraOptions += (findCmdText.match(/BySize/i) ? ' --sz --wt' : ' --wt --sz');
-    extraOptions += ' ' + getOverrideConfigByPriority([folderKey, 'default', ''], 'listSortingFilesOptions') as string || '-l -H 10 -T 10';
+    extraOptions += ' ' + getOverrideConfigByPriority([folderKey, '', 'default'], 'listSortingFilesOptions') as string || '-l -H 10 -T 10';
 
     let searchPathsOptions = getSearchPathOptions(toRunInTerminal, useProjectSpecific, rootFolder, '', FindCommandType.RegexFindAsClassOrMethodDefinitionInCodeFiles === findCmd);
     if (isCookingCmdAlias) {
@@ -196,7 +196,7 @@ export function getFindingCommandByCurrentWord(toRunInTerminal: boolean, findCmd
     let searchPattern = '';
     if (isFindDefinition) {
         searchPattern = MyConfig.UseDefaultFindingClassCheckExtensionRegex.test(parsedFile.ext)
-            ? getConfigValueByRoot(rootFolderName, 'default', '', 'definition')
+            ? getConfigValueByRoot(rootFolderName, '', 'default', 'definition')
             : getConfigValueByRoot(rootFolderName, extension, mappedExt, 'definition');
     } else {
         searchPattern = isFindReference